Revert "Feature/generic alias"
