+++ pybt/typing/basic_types.py
@@ -1,39 +1,52 @@
-from pybt.typing.core import (
-    _BoolGenericAlias,
-    _NoneGenericAlias,
-    BaseType,
-    _FloatGenericAlias,
-    _StringGenericAlias,
-    _IntGenericAlias,
-)
+from pybt.typing.core import _type_check
 
+import random
+import string
+import typing as PythonTyping
 
 """ 
 This file defines pybt types for none, int, str, bool, and float
 """
 
-
-def _type_check(args, types, msg):
-    for idx, el in enumerate(args):
-        if type(el) != types[idx]:
-            raise TypeError(msg)
+_DEFAULT_MIN = -1000
+_DEFAULT_MAX = 1000
+_DEFAULT_MAX_LEN = 10
 
 
-class NoneType(BaseType):
-    _alias = _NoneGenericAlias
+class NoneType:
+    def generate(self) -> None:
+        return None
 
     def __str__(self):
         return "pybt.types.None"
 
-    def __class_getitem__(cls, _):
-        raise TypeError("Expected no argument: NoneType")
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
 
-class Int(BaseType):
-    _alias = _IntGenericAlias
+class Int:
+    def __init__(self, min=_DEFAULT_MIN, max=_DEFAULT_MAX):
+        self.min: int = _DEFAULT_MIN
+        self.max: int = _DEFAULT_MAX
+        if min is not None:
+            self.min = min
+        if max is not None:
+            self.max = max
+
+    def generate(self) -> int:
+        return random.randint(self.min, self.max)
 
     def __str__(self):
         return "pybt.types.Int"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         min = None
@@ -56,14 +69,29 @@ class Int(BaseType):
         if min > max:
             raise TypeError(f"Min {min} is greater than Max {max}")
 
-        return _IntGenericAlias(min, max)
+        return cls(min, max)
 
 
-class Float(BaseType):
-    _alias = _FloatGenericAlias
+class Float:
+    def __init__(self, min=_DEFAULT_MIN, max=_DEFAULT_MAX):
+        self.min: float = _DEFAULT_MIN
+        self.max: float = _DEFAULT_MAX
+        if min is not None:
+            self.min = min
+        if max is not None:
+            self.max = max
+
+    def generate(self) -> float:
+        return random.random() * random.randint(self.min, self.max)
 
     def __str__(self):
         return "pybt.types.Float"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         min = None
@@ -86,14 +114,28 @@ class Float(BaseType):
         if min > max:
             raise TypeError(f"Min {cls.min} is greater than Max {cls.max}")
 
-        return _FloatGenericAlias(min, max)
+        return cls(min, max)
 
 
-class Str(BaseType):
-    _alias = _StringGenericAlias
+class Str:
+    def __init__(self, max_len=_DEFAULT_MAX_LEN):
+        self.max_len: int = _DEFAULT_MAX_LEN
+        if max_len is not None:
+            self.max_len = max_len
+
+    def generate(self) -> str:
+        return "".join(
+            random.choices(string.ascii_letters, k=random.randint(1, self.max_len))
+        )
 
     def __str__(self):
         return "pybt.types.Str"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         max_len = None
@@ -110,14 +152,21 @@ class Str(BaseType):
         if max_len <= 0:
             raise TypeError(f"Max Length of {cls.max_len} is less than or equal 0")
 
-        return _StringGenericAlias(max_len)
+        return cls(max_len)
 
 
-class Bool(BaseType):
-    _alias = _BoolGenericAlias
+class Bool:
+    def generate(self) -> bool:
+        return [True, False][random.randint(0, 1)]
 
     def __str__(self):
         return "pybt.types.Bool"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls):
         raise TypeError("Expected No argument: Bool")
