+++ pybt/typing/complex_types.py
@@ -1,13 +1,9 @@
-from pybt.typing.core import (
-    _AnyGenericAlias,
-    _DictGenericAlias,
-    _FunctionGenericAlias,
-    _ListGenericAlias,
-    _SetGenericAlias,
-    _TupleGenericAlias,
-    _UnionGenericAlias,
-    BaseType,
-)
+import random
+import typing as PythonTyping
+from types import UnionType
+
+from pybt.typing.basic_types import Int, Str, Bool, Float, NoneType
+from pybt.typing.core import _flat_union
 
 
 """ 
@@ -16,12 +12,61 @@ This file implements union, any, list, tuple, dict, function types for PyBT
 TODO: Ensure that types passed are pybt types 
 """
 
+_DEFAULT_MAX_DEPTH = 2
+_DEFAULT_MAX_LEN = 10
+_DEFAULT_SUB_TYPE = None
 
-class Union(BaseType):
-    _alias = _UnionGenericAlias
+
+def _get_next(choices):
+    if not choices:
+        return NoneType()
+    if len(choices) > 1:
+        choice = choices[random.randint(0, len(choices) - 1)]
+    else:
+        choice = choices[0]
+
+    if type(choice) in [
+        Any,
+        List,
+        Tuple,
+        Dict,
+        Int,
+        Str,
+        Bool,
+        Float,
+        NoneType,
+        Set,
+        Function,
+    ]:
+        return choice
+    else:
+        return choice()
+
+
+class Union:
+    def __init__(self, sub_type=_DEFAULT_SUB_TYPE):
+        self.sub_type: UnionType = None
+        if sub_type:
+            self.sub_type = sub_type
+
+    def generate(self) -> any:
+        if not self.sub_type:
+            sub_type_choices = [Any()]
+        else:
+            sub_type_choices = _flat_union(self.sub_type)
+
+        choice = _get_next(sub_type_choices)
+
+        return choice.generate()
 
     def __str__(self):
         return "pybt.types.Union"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         sub_type = None
@@ -33,14 +78,60 @@ class Union(BaseType):
         if len(parameters):
             sub_type = parameters[0]
 
-        return _UnionGenericAlias(sub_type)
+        return cls(sub_type)
 
 
-class Any(BaseType):
-    _alias = _AnyGenericAlias
+class Any:
+    def __init__(self, max_depth=_DEFAULT_MAX_DEPTH, max_len=_DEFAULT_MAX_LEN):
+        self.max_depth: int = _DEFAULT_MAX_DEPTH
+        self.max_len: int = _DEFAULT_MAX_LEN
+        if max_depth is not None:
+            self.max_depth = max_depth
+        if max_len is not None:
+            self.max_len = max_len
+
+    def create_any(self, base_type, times_called):
+        _ALL_TYPES: list = [List, Tuple, Dict, Int, Str, Bool, Float, NoneType]
+        _BASE_TYPES: list = [Int, Str, Bool, Float, NoneType]
+        types = []
+        for _ in range(self.max_len):
+            next = _ALL_TYPES[random.randint(0, len(_ALL_TYPES) - 1)]
+            if next in [List, Tuple, Dict]:
+                if times_called <= self.max_depth:
+                    types.append(self.create_any(next, times_called + 1))
+            else:
+                types.append(next)
+
+        sub_types = PythonTyping._UnionGenericAlias(UnionType, tuple(types))
+
+        if base_type == Dict:
+            key_type = _BASE_TYPES[random.randint(0, len(_BASE_TYPES) - 1)]
+            return base_type[key_type, sub_types]
+
+        return base_type[sub_types]
+
+    def generate(self, base_type=None) -> any:
+        _ALL_TYPES: list = [List, Tuple, Int, Str, Bool, Float, NoneType]
+        _BASE_TYPES: list = [Int, Str, Bool, Float, NoneType]
+        if not base_type:
+            base_type = _ALL_TYPES[random.randint(0, len(_ALL_TYPES) - 1)]
+
+        if base_type in _BASE_TYPES:
+            return base_type().generate()
+        elif type(base_type) in _BASE_TYPES:
+            return base_type.generate()
+
+        type_to_gen = self.create_any(base_type, 0)
+        return type_to_gen.generate()
 
     def __str__(self):
         return "pybt.types.Any"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         max_depth = None
@@ -60,14 +151,36 @@ class Any(BaseType):
         if max_depth <= 0:
             raise TypeError(f"Max Depth of {cls.max_depth} is less than or equal to 0")
 
-        return _AnyGenericAlias(max_len, max_depth)
+        return cls(max_len, max_depth)
 
 
-class List(BaseType):
-    _alias = _ListGenericAlias
+class List:
+    def __init__(self, sub_type=_DEFAULT_SUB_TYPE, max_len=_DEFAULT_MAX_LEN):
+        self.max_len: int = _DEFAULT_MAX_LEN
+        self.sub_type = sub_type
+        if max_len is not None:
+            self.max_len = max_len
+
+    def generate(self) -> list:
+        if not self.sub_type:
+            sub_type_choices = [Any()]
+        else:
+            sub_type_choices = _flat_union(self.sub_type)
+
+        list_gen = []
+        for _ in range(random.randint(0, self.max_len)):
+            t = _get_next(sub_type_choices)
+            list_gen.append(t.generate())
+        return list_gen
 
     def __str__(self):
         return "pybt.types.List"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         sub_type = None
@@ -85,14 +198,28 @@ class List(BaseType):
         if max_len and max_len <= 0:
             raise TypeError(f"Max Length of {cls.max_len} is less than or equal to 0")
 
-        return _ListGenericAlias(sub_type, max_len)
+        return cls(sub_type, max_len)
 
 
-class Tuple(BaseType):
-    _alias = _TupleGenericAlias
+class Tuple:
+    def __init__(self, sub_type=_DEFAULT_SUB_TYPE, max_len=_DEFAULT_MAX_LEN):
+        self.max_len = _DEFAULT_MAX_LEN
+        self.sub_type = sub_type
+        if max_len is not None:
+            self.max_len = max_len
+
+    def generate(self) -> tuple:
+        list_type = List[self.sub_type, self.max_len]
+        return tuple(list_type.generate())
 
     def __str__(self):
         return "pybt.types.Tuple"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         sub_type = None
@@ -110,14 +237,30 @@ class Tuple(BaseType):
         if max_len and max_len <= 0:
             raise TypeError(f"Max Length of {cls.max_len} is less than or equal to 0")
 
-        return _TupleGenericAlias(sub_type, max_len)
+        return cls(sub_type, max_len)
 
 
-class Set(BaseType):
-    _alias = _SetGenericAlias
+class Set:
+    def __init__(self, sub_type=_DEFAULT_SUB_TYPE, max_len=_DEFAULT_MAX_LEN):
+        self.max_len = _DEFAULT_MAX_LEN
+        self.sub_type = sub_type
+        if max_len is not None:
+            self.max_len = max_len
+
+    def generate(self) -> set:
+        if self.sub_type is None: 
+            self.sub_type = Int | Str | Bool | Float | Tuple[Int | Str | Bool | Float]
+        list_type = List[self.sub_type, self.max_len]
+        return set(list_type.generate())
 
     def __str__(self):
         return "pybt.types.Set"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         sub_type = None
@@ -135,14 +278,48 @@ class Set(BaseType):
         if max_len and max_len <= 0:
             raise TypeError(f"Max Length of {cls.max_len} is less than or equal to 0")
 
-        return _SetGenericAlias(sub_type, max_len)
+        return cls(sub_type, max_len)
 
 
-class Dict(BaseType):
-    _alias = _DictGenericAlias
+class Dict:
+    def __init__(
+        self,
+        key_type=_DEFAULT_SUB_TYPE,
+        arg_type=_DEFAULT_SUB_TYPE,
+        max_keys=_DEFAULT_MAX_LEN,
+    ):
+        self.max_keys: int = _DEFAULT_MAX_LEN
+        self.key_type = key_type
+        self.arg_type = arg_type
+        if max_keys is not None:
+            self.max_keys = max_keys
+
+    def generate(self) -> dict:
+        if not self.key_type:
+            key_type_choices = [Int, Float, Bool, Str]
+        else:
+            key_type_choices = _flat_union(self.key_type)
+
+        if not self.arg_type:
+            arg_type_choices = [Any()]
+        else:
+            arg_type_choices = _flat_union(self.arg_type)
+
+        dict_gen = {}
+        for _ in range(random.randint(0, self.max_keys)):
+            key = _get_next(key_type_choices)
+            arg = _get_next(arg_type_choices)
+            dict_gen[key.generate()] = arg.generate()
+        return dict_gen
 
     def __str__(self):
         return "pybt.types.Dict"
+
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
 
     def __class_getitem__(cls, parameters):
         key_type = None
@@ -166,16 +343,33 @@ class Dict(BaseType):
         if max_keys and max_keys <= 0:
             raise TypeError(f"Max # of Keys of {cls.max_keys} is less than or equal 0")
 
-        return _DictGenericAlias(key_type, arg_type, max_keys)
+        return cls(key_type, arg_type, max_keys)
 
 
-class Function(BaseType):
-    _alias = _FunctionGenericAlias
+class Function:
+    def __init__(self, return_type=None):
+        self.return_type = return_type
+
+    def generate(self):
+        if not self.return_type:
+            sub_type_choices = [Any()]
+        else:
+            sub_type_choices = _flat_union(self.return_type)
+
+        t = _get_next(sub_type_choices)
+        return lambda *x: t.generate()
 
     def __str__(self):
         return "pybt.types.Function"
 
+    def __or__(self, other):
+        return PythonTyping.Union[self, other]
+
+    def __ror__(self, other):
+        return PythonTyping.Union[self, other]
+
     def __class_getitem__(cls, parameters):
+        # Function annotates only its return type
         return_type = None
         if type(parameters) != tuple:
             parameters = (parameters,)
@@ -185,4 +379,6 @@ class Function(BaseType):
         if len(parameters):
             return_type = parameters[0]
 
-        return _FunctionGenericAlias(return_type)
+        return cls(return_type)
+
+    # Function[[], ...]
